Code clean: modularize server/main.js, extracting mongoDB and app init functions
